+++ package.json
@@ -14,6 +14,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@supabase/supabase-js": "^2.42.0",
     "firebase": "^9.23.0",
     "framer-motion": "^4.1.17",
     "gsap": "^3.12.5",
