+++ pages/projects/[pid].js
